# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ETK API",
        "contact": {
            "name": "fbeck",
            "email": "loordy@live.ru"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        },
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        }
    ],
    "paths": {
...
        "/reports/basebysoato": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "operationId": "/reports/basebysoato",
                "parameters": [
                    {
                        "name": "soato",
                        "in": "query",
                        "description": "СОАТО региона",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistic retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "properties": {
                                                    "soato": {
                                                        "type": "string",
                                                        "example": "1726"
                                                    },
                                                    "company_count": {
                                                        "type": "string",
                                                        "example": "10"
                                                    },
                                                    "workplaces": {
                                                        "type": "string",
                                                        "example": "54"
                                                    },
                                                    "vacancies": {
                                                        "type": "string",
                                                        "example": "17"
                                                    },
                                                    "total_rates": {
                                                        "type": "string",
                                                        "example": "48.75"
                                                    },
                                                    "vacant_rates": {
                                                        "type": "string",
                                                        "example": "17.00"
                                                    },
                                                    "employed": {
                                                        "type": "string",
                                                        "example": "37"
                                                    },
                                                    "active_contracts_in_gender_male": {
                                                        "type": "string",
                                                        "example": "32"
                                                    },
                                                    "active_contracts_in_gender_female": {
                                                        "type": "string",
                                                        "example": "5"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Statistic retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "etk_auth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}